move struts 2.3 and 7 instrumenation to separate pacakges
