+++ instrumentation/struts/struts-2.3/javaagent/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
   testInstrumentation(project(":instrumentation:servlet:servlet-3.0:javaagent"))
   testInstrumentation(project(":instrumentation:servlet:servlet-javax-common:javaagent"))
   testInstrumentation(project(":instrumentation:jetty:jetty-8.0:javaagent"))
+  testInstrumentation(project(":instrumentation:struts:struts-7.0:javaagent"))
 
   latestDepTestLibrary("org.apache.struts:struts2-core:6.0.+")
 }
