+++ instrumentation/struts/struts-2.3/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/struts2/v2_3/ActionInvocationInstrumentation.java
@@ -3,16 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.struts2;
+package io.opentelemetry.javaagent.instrumentation.struts2.v2_3;
 
 import static io.opentelemetry.instrumentation.api.semconv.http.HttpServerRouteSource.CONTROLLER;
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.hasClassesNamed;
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.implementsInterface;
-import static io.opentelemetry.javaagent.instrumentation.struts2.StrutsSingletons.instrumenter;
+import static io.opentelemetry.javaagent.instrumentation.struts2.v2_3.StrutsSingletons.instrumenter;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
 
 import com.opensymphony.xwork2.ActionInvocation;
 import io.opentelemetry.context.Context;
@@ -29,8 +28,7 @@ public class ActionInvocationInstrumentation implements TypeInstrumentation {
 
   @Override
   public ElementMatcher<ClassLoader> classLoaderOptimization() {
-    return hasClassesNamed("com.opensymphony.xwork2.ActionInvocation")
-        .and(not(hasClassesNamed("org.apache.struts2.ActionInvocation")));
+    return hasClassesNamed("com.opensymphony.xwork2.ActionInvocation");
   }
 
   @Override
