+++ instrumentation/struts/struts-2.3/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/struts2/v2_3/StrutsCodeAttributesGetter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.struts2;
+package io.opentelemetry.javaagent.instrumentation.struts2.v2_3;
 
 import com.opensymphony.xwork2.ActionInvocation;
 import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesGetter;
