+++ instrumentation/struts/struts-7.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/struts2/v7_0/GreetingAction.java
@@ -3,15 +3,16 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.struts2;
+package io.opentelemetry.javaagent.instrumentation.struts2.v7_0;
 
 import static io.opentelemetry.instrumentation.testing.junit.http.AbstractHttpServerTest.controller;
 
-import com.opensymphony.xwork2.ActionSupport;
 import io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.apache.struts2.ActionSupport;
 import org.apache.struts2.ServletActionContext;
+import org.apache.struts2.interceptor.parameter.StrutsParameter;
 
 public class GreetingAction extends ActionSupport {
 
@@ -80,6 +81,7 @@ public class GreetingAction extends ActionSupport {
     return "greetingServlet";
   }
 
+  @StrutsParameter
   public void setId(String id) {
     responseBody = id;
   }
